Rename lava_mock.rs to server.rs
This is a separate commit because it significantly improves git's
ability to reason about the rename.
