+++ lava-api-mock/src/lib.rs
@@ -69,7 +69,7 @@
 mod devices;
 mod devicetypes;
 mod jobs;
-mod lava_mock;
+mod server;
 mod state;
 mod tags;
 mod testcases;
@@ -80,7 +80,7 @@ pub use devices::{Device, Health as DeviceHealth, State as DeviceState};
 pub use devicetypes::{Alias, Architecture, BitWidth, Core, DeviceType, ProcessorFamily};
 pub use jobs::Job;
 pub use jobs::{Health as JobHealth, State as JobState};
-pub use lava_mock::{PaginationLimits, Server};
+pub use server::{PaginationLimits, Server};
 pub use state::{PopulationParams, SharedState, State};
 pub use tags::Tag;
 pub use testcases::{Metadata, PassFail, TestCase, TestSet, TestSuite};
